Decomp func_80074950

+++ src/BATTLE/BATTLE.PRG/4A0A8.h
@@ -3,7 +3,7 @@
 void func_800B28A8(void*, void*, char);
 int func_800BEB10(void);
 void func_800BEBEC(void);
-char func_800BEBF4(int);
+char func_800BEBF4(short);
 void func_800BEC14(short, int);
 void func_800BEC30(void);
 short func_800BEC58(int, int, char*, int);
